MM-15694: Illustrate plugin.DismissPostError (#50)
Tweak the `MessageWillBePosted` hook to dismiss posts that try to
at-mention the `@demo_plugin` user.

When enabled, the plugin already disallows posts in the demo plugin channel, but does so by displaying an error and causing the post to enter a retry/cancel state. It used to do that for at-mentions in any channel, but is now changed as per above.